+++ mne/source_space.py
@@ -205,10 +205,10 @@ class SourceSpaces(list):
 
     @property
     def kind(self):
-        """The kind of source space (surface, volume, discrete)."""
+        """The kind of source space (surface, volume, discrete, mixed)."""
         ss_types = list(set([ss['type'] for ss in self]))
         if len(ss_types) != 1:
-            return 'combined'
+            return 'mixed'
         return _src_kind_dict[ss_types[0]]
 
     def __add__(self, other):